fastapi-demo/pyproject.toml
[project]
name = 'fastapi-demo'
version = '1.0.0'
requires-python = ">=3.13"

dependencies = [
    "fastapi>=0.121.1",
    "fastapi-admin>=1.0.4",
    "tortoise-orm[aiosqlite]>=0.25.1",
]
